app/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
blinker==1.8.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Django==4.2.17
Flask==3.0.3
fonttools==4.55.3
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
plotly==5.24.1
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
squarify==0.4.4
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.6
zipp==3.20.2
